Fix small issues in ABC-SMC (#345)
* fix output names

SMC did not work with additional output names. this was because additional output names were interpreted as the full output names list and essential variables were not recorded.

* fix mismatched dimension

previous version resulted in mismatch between dimensions when mean vectors are 1-dimensional. this was due to an extra dimension which is now removed in parameter normalisation.

* update changelog

* Upgrade dask explicitly for travis

Co-authored-by: hpesonen <henri.e.pesonen@gmail.com>