fix nested acceptance
about nested acceptance: rejection sampler sorts samples based on distance and assumes that the n_samples first samples are the accepted sample. however when several distance functions and and thresholds are used as acceptance criterion, the smallest samples selected based on any one distance measure could contain samples that are not acceptable based on another distance measure and associated threshold.

about what was fixed: previous version checked whether a new batch contains samples where one or more distance measure is not below threshold and substituted all distances in such samples with infinities. this would ensure that the samples would not be sorted as smallest ones when the batch is merged with previous samples. however based on travis, the distances that were overwritten are a read-only variable in linux. hence this new version does not overwrite the distances but checks whether there are acceptable samples in a batch and if so, merges these with the previous samples.