Add adaptive distance SMC-ABC (#346)
* add adaptive distance SMC

note: uses standard rejection sampler

* add adaptive distance node

added new class and updated adaptive distance SMC to use the adaptive distance node

* add nested acceptance

updates:
- adaptive distance node stores and evaluates all distance functions
- rejection sampler determines acceptance based on comparison between all distances and thresholds

note: removed adaptive rejection sampler

* add functional test

* allow user to continue iterative estimation

since adaptive distance SMC behaviour in each new round is determined based on previous populations, it is possible to let users continue iteration estimation.

* adaptive distance SMC without pool

this is an alternative implementation that does not use elfi output pool to store observations needed in distance adaptation but uses rejection sampler to handle adaptation.

* fix first iteration threshold

adaptive distance SMC did not work when batch size is smaller than the requested sample size because rejection sampler would generate the batch size rather than requested sample size when called with an infinite threshold.

* add consistency test

* fix stored statistics calculation

the adaptive distance node uses observed summaries data to calculate statistics based on which standard deviation can be estimated. previous version assumed that the individual summaries are scalar and did not work with vector summaries.

* add docstrings

* fix weighted distance calculation

* add unit tests

- check that the scale in adaptive distance node is calculated based on all data added between updates
- check that the distances in adaptive distance node correspond to euclidean distance between scaled summaries

* revise standard deviation update

previous version calculated standard deviation based on running totals that allowed incremental update but could lead to overflow problems.

* fix adaptive distance initialisation

* fix style issues

* fix docstrings style

* update changelog

* update method name

* fix nested acceptance

about nested acceptance: rejection sampler sorts samples based on distance and assumes that the n_samples first samples are the accepted sample. however when several distance functions and and thresholds are used as acceptance criterion, the smallest samples selected based on any one distance measure could contain samples that are not acceptable based on another distance measure and associated threshold.

about what was fixed: previous version checked whether a new batch contains samples where one or more distance measure is not below threshold and substituted all distances in such samples with infinities. this would ensure that the samples would not be sorted as smallest ones when the batch is merged with previous samples. however based on travis, the distances that were overwritten are a read-only variable in linux. hence this new version does not overwrite the distances but checks whether there are acceptable samples in a batch and if so, merges these with the previous samples.

* update docs

* update readme

Co-authored-by: Henri Pesonen <henri.e.pesonen@gmail.com>